refactor: 💡 more resource filtering name changes
