+++ addons/core/addon/helpers/has-route-resource-filter-selections.js
@@ -4,18 +4,18 @@ import { getOwner } from '@ember/application';
 import { action } from '@ember/object';
 
 /**
- * Use this helper to fetch the current value of a resource filter param,
- * as well as its allowedValues.
+ * Use this helper to determine if a resource filter for a route has any
+ * currently selected value.
  *
  * @see resourceFilterParam (decorator)
  *
  * @example
  *
- *   {{#with resource-filter-param 'status' as |param|}}
- *     {{param.name}} / {{param.value}} / {{param.selectedValue}}
- *   {{/with}}
+ *   {{#if (has-route-resource-filter-selections 'route.name' 'field')}}
+ *     Field has filter selections.
+ *   {{/if}}
  */
-export default class ResourceFilterParamHelper extends Helper {
+export default class HasRouteResourceFilterSelectionsHelper extends Helper {
   // =services
 
   @service router;
@@ -43,17 +43,21 @@ export default class ResourceFilterParamHelper extends Helper {
   // =compute method
 
   /**
-   * Returns an object containing `name`, `value`, and `selectedValue`.
-   * `name` - The name of the resource filter param.
-   * `value` - The value of the resource filter param
-   * @return {object}
+   * Returns true if the specified route has a value set for
+   * any resource filter passed into the helper.
+   * @param {string} routeName
+   * @param {string[]} names - names of resource filter fields
+   * @return {boolean}
    */
-  compute([routeName, name]) {
+  compute([routeName, ...names]) {
     const owner = getOwner(this);
-    const containerKey = `route-resource-filter:${name}@${routeName}`;
-    const resourceFilter = owner.lookup(containerKey);
-    const { allowedValues, value: selectedValue } = resourceFilter;
-    return { name, allowedValues, selectedValue };
+    const route = owner.lookup(`route:${routeName}`);
+    const selectedValues = names.map((name) => route[name] || null).flat();
+    const anyTruthy = selectedValues.reduce(
+      (previousValue, currentValue) => previousValue || currentValue,
+      false
+    );
+    return Boolean(anyTruthy);
   }
 
   // =actions
