--- addons/core/app/decorators/resource-filter-param.js
@@ -1,1 +1,0 @@
-export { default } from 'core/decorators/resource-filter-param';
