+++ addons/core/app/decorators/route-resource-filter.js
@@ -1,0 +1,1 @@
+export { default } from 'core/decorators/route-resource-filter';
