--- addons/core/app/helpers/has-resource-filter-param-selections.js
@@ -1,1 +1,0 @@
-export { default } from 'core/helpers/has-resource-filter-param-selections';
