+++ addons/core/app/helpers/has-route-resource-filter-selections.js
@@ -1,0 +1,1 @@
+export { default } from 'core/helpers/has-route-resource-filter-selections';
