--- addons/core/app/helpers/resource-filter-param.js
@@ -1,1 +1,0 @@
-export { default } from 'core/helpers/resource-filter-param';
