+++ addons/core/app/helpers/route-resource-filter.js
@@ -1,0 +1,1 @@
+export { default } from 'core/helpers/route-resource-filter';
