+++ ui/admin/app/routes/scopes/scope/auth-methods.js
@@ -4,7 +4,7 @@ import { action } from '@ember/object';
 import loading from 'ember-loading/decorator';
 import { confirm } from 'core/decorators/confirm';
 import { notifySuccess, notifyError } from 'core/decorators/notify';
-import { resourceFilterParam } from 'core/decorators/resource-filter-param';
+import { resourceFilterParam } from 'core/decorators/route-resource-filter';
 
 export default class ScopesScopeAuthMethodsRoute extends Route {
   // =services
